Zombies 0.16b
fixed the sound button - for now there is a sound button in the game
that you can press to toggle sounds on/off. Chris is looking into a way
to put it into the menuState, but we may end up just having to make it
an in game option menu.